% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_acl_get}
\alias{aws_bucket_acl_get}
\title{Get a bucket ACL}
\usage{
aws_bucket_acl_get(bucket)
}
\arguments{
\item{bucket}{(character) bucket name. required}
}
\value{
xxx
}
\description{
Get a bucket ACL
}
\details{
see docs at \url{https://www.paws-r-sdk.com/docs/s3_get_bucket_acl/}
}
\examples{
\dontrun{
aws_bucket_acl_get("s3://s64-test-2")
}
}
