% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{aws_bucket_add_user}
\alias{aws_bucket_add_user}
\title{Add a user to a bucket}
\usage{
aws_bucket_add_user(bucket, username, permissions)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{username}{(character) A user name. required}

\item{permissions}{(character) user permissions, one of
read or write. write includes read}
}
\value{
invisibly returns nothing
}
\description{
Add a user to a bucket
}
\section{Permissions}{

\itemize{
\item read: read only; not allowed to write or do admin tasks
\item write: write (in addition to read); includes deleting files; does
not include deleting buckets
\item admin: change user permissions (in addition to read and write);
includes deleting buckets
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_bucket_add_user(
  bucket = "s64-test-22",
  username = "scott",
  permissions = "read"
)
\dontrun{
# not a valid permissions string
aws_bucket_add_user(
  bucket = "mybucket",
  username = "scott",
  permissions = "notavalidpermission"
)
}
\dontshow{\}) # examplesIf}
}
