% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{aws_bucket_get_permissions}
\alias{aws_bucket_get_permissions}
\title{Get permissions for a bucket}
\usage{
aws_bucket_get_permissions(bucket)
}
\arguments{
\item{bucket}{(character) bucket name. required}
}
\value{
tibble with a row for each user, with columns:
\itemize{
\item user
\item permissions
}

Note that users with no persmissions are not shown; see \code{\link[=aws_users]{aws_users()}}
}
\description{
Get permissions for a bucket
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_bucket_get_permissions("s64-test-22")
\dontshow{\}) # examplesIf}
}
