% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{aws_bucket_remove_user}
\alias{aws_bucket_remove_user}
\title{Remove a user from a bucket}
\usage{
aws_bucket_remove_user(bucket, username)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{username}{(character) A user name. required}
}
\value{
invisibly returns nothing
}
\description{
Remove a user from a bucket
}
\details{
This function detaches a policy from a user for accessing
the bucket; the policy itself is untouched
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_bucket_remove_user("s64-test-22", "scott")
\dontshow{\}) # examplesIf}
}
