% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{aws_configure}
\alias{aws_configure}
\title{Configure sixtyfour settings}
\usage{
aws_configure(redacted = FALSE, redact_str = "*****", verbose = TRUE)
}
\arguments{
\item{redacted}{(logical) Redact secrets? Default: \code{FALSE}. If \code{TRUE},
secret values are redacted (replaced with \verb{*****}) only in printed
messages to the console, and are not redacted in the output from
running functions.}

\item{redact_str}{(character) String to use to replace redacted values.
Default: \code{"*****"}}

\item{verbose}{(logical) Print verbose output? Default: \code{TRUE}.}
}
\description{
Configure sixtyfour settings
}
