% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-redshift.R
\name{aws_db_cluster_status}
\alias{aws_db_cluster_status}
\title{Get cluster status}
\usage{
aws_db_cluster_status(id)
}
\arguments{
\item{id}{(character) Cluster identifier. Use this identifier to refer to
the cluster for any subsequent cluster operations such as deleting or
modifying. The identifier also appears in the Amazon Redshift console.
Must be unique for all clusters within a Amazon Web Services account.}
}
\value{
(character) the status of the cluster, e.g., "creating",
"available", "not found"
}
\description{
Get cluster status
}
\examples{
\dontrun{
aws_db_cluster_status(id = "scotts-test-cluster-456")
}
}
\seealso{
Other database:
\code{\link[=aws_db_instance_status]{aws_db_instance_status()}},
\code{\link[=aws_db_rds_con]{aws_db_rds_con()}},
\code{\link[=aws_db_rds_create]{aws_db_rds_create()}},
\code{\link[=aws_db_rds_list]{aws_db_rds_list()}},
\code{\link[=aws_db_redshift_con]{aws_db_redshift_con()}},
\code{\link[=aws_db_redshift_create]{aws_db_redshift_create()}}
}
\concept{database}
