% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-rds.R
\name{aws_db_instance_details}
\alias{aws_db_instance_details}
\title{Get information for all RDS instances}
\usage{
aws_db_instance_details()
}
\value{
a list of RDS instance details, see link below for format, with slots:
\itemize{
\item Marker (for pagination)
\item DBInstances (each instance; empty list if no instances)
}
}
\description{
Get information for all RDS instances
}
\references{
\url{https://www.paws-r-sdk.com/docs/describe_db_instances/}
}
