% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-rds.R
\name{aws_db_instance_status}
\alias{aws_db_instance_status}
\title{Get instance status}
\usage{
aws_db_instance_status(id)
}
\arguments{
\item{id}{(character) required. instance identifier. The identifier for
this DB instance. This parameter is stored as a lowercase string.
Constraints: must contain from 1 to 63 letters, numbers, or hyphens; first
character must be a letter; can't end with a hyphen or contain two
consecutive hyphens. required.}
}
\value{
(character) the status of the instance, e.g., "creating",
"available", "not found"
}
\description{
Get instance status
}
\examples{
\dontrun{
aws_db_instance_status(id = "thedbinstance")
}
}
\seealso{
Other database:
\code{\link[=aws_db_cluster_status]{aws_db_cluster_status()}},
\code{\link[=aws_db_rds_con]{aws_db_rds_con()}},
\code{\link[=aws_db_rds_create]{aws_db_rds_create()}},
\code{\link[=aws_db_rds_list]{aws_db_rds_list()}},
\code{\link[=aws_db_redshift_con]{aws_db_redshift_con()}},
\code{\link[=aws_db_redshift_create]{aws_db_redshift_create()}}
}
\concept{database}
