% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-rds.R
\name{aws_db_rds_list}
\alias{aws_db_rds_list}
\title{Get information for all RDS instances}
\usage{
aws_db_rds_list()
}
\value{
a tibble of instance details;
see \url{https://www.paws-r-sdk.com/docs/rds_describe_db_instances/}
an empty tibble if no instances found
}
\description{
Get information for all RDS instances
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
aws_db_rds_list()
\dontshow{\}) # examplesIf}
}
\seealso{
Other database:
\code{\link[=aws_db_cluster_status]{aws_db_cluster_status()}},
\code{\link[=aws_db_instance_status]{aws_db_instance_status()}},
\code{\link[=aws_db_rds_con]{aws_db_rds_con()}},
\code{\link[=aws_db_rds_create]{aws_db_rds_create()}},
\code{\link[=aws_db_redshift_con]{aws_db_redshift_con()}},
\code{\link[=aws_db_redshift_create]{aws_db_redshift_create()}}
}
\concept{database}
