% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{aws_db_redshift_client}
\alias{aws_db_redshift_client}
\title{Get the \code{paws} Redshift client}
\usage{
aws_db_redshift_client()
}
\value{
a list with methods for interfacing with Redshift;
see \url{https://www.paws-r-sdk.com/docs/redshift/}
}
\description{
Get the \code{paws} Redshift client
}
