% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{aws_ec2_client}
\alias{aws_ec2_client}
\title{Get the \code{paws} EC2 client - primarily for usage of VPC security groups}
\usage{
aws_ec2_client()
}
\value{
a list with methods for interfacing with EC2;
see \url{https://www.paws-r-sdk.com/docs/ec2/}
}
\description{
Get the \code{paws} EC2 client - primarily for usage of VPC security groups
}
\note{
returns existing client if found; a new client otherwise
}
\concept{security groups}
