% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{aws_groups}
\alias{aws_groups}
\title{List all groups or groups for a single user}
\usage{
aws_groups(username = NULL, ...)
}
\arguments{
\item{username}{(character) a username. optional}

\item{...}{parameters passed on to \code{paws} \code{list_groups_for_user}
if username is non-NULL, otherwise passed on to \code{list_users}}
}
\value{
A tibble with information about groups
}
\description{
List all groups or groups for a single user
}
\examples{
\dontshow{if (aws_has_creds()) withAutoprint(\{ # examplesIf}
aws_groups()
aws_groups(username = aws_user_current())
\dontshow{\}) # examplesIf}
}
\seealso{
Other groups:
\code{\link[=aws_group]{aws_group()}},
\code{\link[=aws_group_create]{aws_group_create()}},
\code{\link[=aws_group_delete]{aws_group_delete()}},
\code{\link[=aws_group_exists]{aws_group_exists()}},
\code{\link[=six_group_delete]{six_group_delete()}}
}
\concept{groups}
