% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{aws_groups_list}
\alias{aws_groups_list}
\title{List groups}
\usage{
aws_groups_list(...)
}
\arguments{
\item{...}{parameters passed on to the \code{paws}
\href{https://www.paws-r-sdk.com/docs/iam_list_groups/}{list_users} method}
}
\value{
A tibble with information about groups
}
\description{
List groups
}
\examples{
\dontrun{
aws_groups_list()
}
}
