% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aws_has_creds}
\alias{aws_has_creds}
\title{Check if appropriate AWS credentials are available}
\usage{
aws_has_creds()
}
\description{
Check if appropriate AWS credentials are available
}
\examples{
aws_has_creds()
}
