% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam.R
\name{aws_iam_client}
\alias{aws_iam_client}
\title{Get the \code{paws} Identity and Access Management (IAM) client}
\usage{
aws_iam_client()
}
\value{
a list with methods for interfacing with IAM;
see \url{https://www.paws-r-sdk.com/docs/iam/}
}
\description{
Get the \code{paws} Identity and Access Management (IAM) client
}
\examples{
aws_iam_client()
}
