% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{aws_policies_attach}
\alias{aws_policies_attach}
\title{Attach more than one policy to a user, group, or role}
\usage{
aws_policies_attach(.x, policies)
}
\arguments{
\item{.x}{result of a call to create or get method for user,
group, or role}

\item{policies}{(character) vector of policies}
}
\value{
NULL, invisibly
}
\description{
Attach more than one policy to a user, group, or role
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_group("read") \%>\% aws_policies_attach(aws_policy_set("read"))
\dontshow{\}) # examplesIf}
}
