% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{aws_policies_list}
\alias{aws_policies_list}
\title{List policies}
\usage{
aws_policies_list(...)
}
\arguments{
\item{...}{parameters passed on to the \code{paws}
\href{https://www.paws-r-sdk.com/docs/iam_list_policies/}{list_users} method}
}
\value{
A tibble with information about policies
}
\description{
List policies
}
\examples{
\dontrun{
aws_policies_list()
}
}
