% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{aws_policy_set}
\alias{aws_policy_set}
\title{Get a set of policies for either read or write}
\usage{
aws_policy_set(set)
}
\arguments{
\item{set}{(character) one of "read" or "write"}
}
\value{
character vector of AWS managed policy names
}
\description{
Get a set of policies for either read or write
}
\examples{
aws_policy_set("read")
aws_policy_set("write")
# aws_policy_set("apple")
}
