% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{aws_role_exists}
\alias{aws_role_exists}
\title{Check if a role exists}
\usage{
aws_role_exists(name)
}
\arguments{
\item{name}{(character) the role name}
}
\value{
a single boolean
}
\description{
Check if a role exists
}
\examples{
\dontshow{if (aws_has_creds()) withAutoprint(\{ # examplesIf}
aws_role_exists("AWSServiceRoleForRedshift")
aws_role_exists("NotARole")
\dontshow{\}) # examplesIf}
}
\seealso{
Other roles:
\code{\link[=aws_role]{aws_role()}},
\code{\link[=aws_role_create]{aws_role_create()}},
\code{\link[=aws_role_delete]{aws_role_delete()}},
\code{\link[=aws_roles]{aws_roles()}}
}
\concept{roles}
