% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{aws_secrets_all}
\alias{aws_secrets_all}
\title{Get all secret values}
\usage{
aws_secrets_all()
}
\value{
(tbl) with secrets
}
\description{
Get all secret values
}
\examples{
\dontshow{if (aws_has_creds() && interactive()) withAutoprint(\{ # examplesIf}
aws_secrets_all()
\dontshow{\}) # examplesIf}
}
