% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{aws_secrets_list}
\alias{aws_secrets_list}
\title{List secrets}
\usage{
aws_secrets_list(...)
}
\arguments{
\item{...}{parameters passed on to the \code{paws} method}
}
\value{
(list) list with secrets
}
\description{
List secrets
}
\note{
see \url{https://www.paws-r-sdk.com/docs/secretsmanager_list_secrets/}
for available parameters
}
\examples{
\dontshow{if (aws_has_creds() && interactive()) withAutoprint(\{ # examplesIf}
aws_secrets_list()
\dontshow{\}) # examplesIf}
}
