% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_add_to_rds}
\alias{aws_user_add_to_rds}
\title{Grant a user access to an RDS database}
\usage{
aws_user_add_to_rds(user, id)
}
\arguments{
\item{user}{(character) an IAM username. required}

\item{id}{(character) instance identifier. required}
}
\value{
xxx
}
\description{
Grant a user access to an RDS database
}
\examples{
\dontrun{
aws_user_add_to_rds(user = "sean", id = "bluebird")
}
}
