% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{aws_user_creds}
\alias{aws_user_creds}
\title{Create access keys for a user}
\usage{
aws_user_creds(username, copy_to_cp = FALSE)
}
\arguments{
\item{username}{(character) A user name. required}

\item{copy_to_cp}{(logical) Copy to clipboard. Default: \code{FALSE}. See
section "Clipboard" below for more details.}
}
\value{
NULL invisibly
}
\description{
Creates a new Amazon Web Services secret access key and
corresponding Amazon Web Services access key ID
}
\details{
A user can have more than one pair of access keys.
By default a user can have up to 2 pairs of access keys.
Using this function will not replace an existing set of keys;
but instead adds an additional set of keys.

See \url{https://rstats.wtf/r-startup.html} for help on bringing in secrets
to an R session.
}
\section{Important}{

Save the secret key after running this function as it can not be
viewed again.
}

\section{Clipboard}{

If you set \code{copy_to_cp=TRUE} we'll copy to your clipboard an
email template with the credentials and a small amount of instructions.
Please do edit that email with information tailored to your
group and how you'd like to store secrets.
}

\section{Known error behaviors}{

\itemize{
\item \verb{LimitExceeded (HTTP 409). Cannot exceed quota for AccessKeysPerUser: 2}
\item \verb{NoSuchEntity (HTTP 404). The user with name xxx cannot be found.}
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
if (!aws_user_exists("jane")) aws_user_create("jane")
aws_user_creds("jane")
aws_user_access_key("jane")
aws_user_creds("jane", copy_to_cp = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aws_user_access_key]{aws_user_access_key()}}, \code{\link[=aws_user_access_key_delete]{aws_user_access_key_delete()}}
}
