% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_users}
\alias{aws_users}
\title{List Users}
\usage{
aws_users(...)
}
\arguments{
\item{...}{parameters passed on to the \code{paws}
\href{https://www.paws-r-sdk.com/docs/iam_list_users/}{list_users} method}
}
\value{
A tibble with information about user accounts, with columns:
\itemize{
\item UserName
\item UserId
\item Path
\item Arn
\item CreateDate
\item PasswordLastUsed
}
}
\description{
List Users
}
\examples{
\dontshow{if (interactive() && aws_has_creds()) withAutoprint(\{ # examplesIf}
aws_users()
\dontshow{\}) # examplesIf}
}
\seealso{
Other users:
\code{\link[=aws_user]{aws_user()}},
\code{\link[=aws_user_access_key]{aws_user_access_key()}},
\code{\link[=aws_user_access_key_delete]{aws_user_access_key_delete()}},
\code{\link[=aws_user_add_to_group]{aws_user_add_to_group()}},
\code{\link[=aws_user_create]{aws_user_create()}},
\code{\link[=aws_user_current]{aws_user_current()}},
\code{\link[=aws_user_delete]{aws_user_delete()}},
\code{\link[=aws_user_exists]{aws_user_exists()}},
\code{\link[=six_user_create]{six_user_create()}},
\code{\link[=six_user_delete]{six_user_delete()}}
}
\concept{users}
