% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcs.R
\name{aws_vpc}
\alias{aws_vpc}
\title{Get a VPC by id}
\usage{
aws_vpc(id, ...)
}
\arguments{
\item{id}{(character) The id of the VPC. required}

\item{...}{parameters passed on to \href{https://www.paws-r-sdk.com/docs/ec2_describe_vpcs/}{describe_vpcs}}
}
\value{
(list) with fields:
\itemize{
\item Vpcs (list) each VPC group
\item NextToken (character) token for paginating
}

Each element of Vpcs is a list with slots:
\itemize{
\item CidrBlock
\item DhcpOptionsId
\item State
\item VpcId
\item OwnerId
\item InstanceTenancy
\item Ipv6CidrBlockAssociationSet
\item CidrBlockAssociationSet
\item IsDefault
\item Tags
}
}
\description{
Get a VPC by id
}
