% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{aws_vpc_sec_group_rules}
\alias{aws_vpc_sec_group_rules}
\title{Modify security group rules}
\usage{
aws_vpc_sec_group_rules(id, rules, ...)
}
\arguments{
\item{id}{(character) security group id}

\item{rules}{list of rules to add/modify on the security group \code{id}}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/ec2_modify_security_group_rules/}{modify_security_group_rules}}
}
\description{
Modify security group rules
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_vpc_sec_group_rules(
  id = "someid",
  rules = list(
    SecurityGroupRuleId = "sgr-07de36a0521f39c8b",
    SecurityGroupRule = list(
      IpProtocol = "tcp",
      FromPort = 22,
      ToPort = 22,
      CidrIpv4 = "3.3.3.3/32",
      Description = "added ssh port"
    )
  )
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other security groups: 
\code{\link{aws_vpc_security_group_create}()},
\code{\link{aws_vpc_security_group_ingress}()},
\code{\link{aws_vpc_security_groups}()},
\code{\link{aws_vpc_security_group}()},
\code{\link{aws_vpc_sg_with_ingress}()}
}
\concept{security groups}
