% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{aws_vpc_security_group_ingress}
\alias{aws_vpc_security_group_ingress}
\title{Authorize Security Group Ingress}
\usage{
aws_vpc_security_group_ingress(id, ip_permissions = NULL, ...)
}
\arguments{
\item{id}{(character) security group id. required}

\item{ip_permissions}{(list) list of persmissions. see link to \code{paws}
docs below or use \code{\link[=ip_permissions_generator]{ip_permissions_generator()}} to generate the
list for this parameter}

\item{...}{named parameters passed on to
\href{https://www.paws-r-sdk.com/docs/ec2_authorize_security_group_ingress/}{authorize_security_group_ingress}}
}
\value{
list with slots:
\itemize{
\item Return (boolean)
\item SecurityGroupRules (list)
\itemize{
\item SecurityGroupRuleId
\item GroupId
\item GroupOwnerId
\item IsEgress
\item IpProtocol
\item FromPort
\item ToPort
\item CidrIpv4
\item CidrIpv6
\item PrefixListId
\item ReferencedGroupInfo
\item Description
\item Tags
}
}
}
\description{
Authorize Security Group Ingress
}
\seealso{
Other security groups:
\code{\link[=aws_vpc_sec_group_rules_mod]{aws_vpc_sec_group_rules_mod()}},
\code{\link[=aws_vpc_security_group]{aws_vpc_security_group()}},
\code{\link[=aws_vpc_security_group_create]{aws_vpc_security_group_create()}},
\code{\link[=aws_vpc_security_groups]{aws_vpc_security_groups()}},
\code{\link[=aws_vpc_sg_with_ingress]{aws_vpc_sg_with_ingress()}}
}
\concept{security groups}
