% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{aws_vpc_sg_with_ingress}
\alias{aws_vpc_sg_with_ingress}
\title{Get a security group with one ingress rule based on the engine}
\usage{
aws_vpc_sg_with_ingress(engine)
}
\arguments{
\item{engine}{(character) The engine to use. default: "mariadb". required.
one of: mariadb, mysql, postgres, or redshift}
}
\value{
(character) security group ID
}
\description{
Get a security group with one ingress rule based on the engine
}
\details{
Adds an ingress rule specific to the \code{engine} supplied (port
changes based on the engine), and your IP address. To create your own
security group and ingress rules see \code{\link[=aws_vpc_security_group_create]{aws_vpc_security_group_create()}}
and \code{\link[=aws_vpc_security_group_ingress]{aws_vpc_security_group_ingress()}}
}
\seealso{
Other security groups:
\code{\link[=aws_vpc_sec_group_rules_mod]{aws_vpc_sec_group_rules_mod()}},
\code{\link[=aws_vpc_security_group]{aws_vpc_security_group()}},
\code{\link[=aws_vpc_security_group_create]{aws_vpc_security_group_create()}},
\code{\link[=aws_vpc_security_group_ingress]{aws_vpc_security_group_ingress()}},
\code{\link[=aws_vpc_security_groups]{aws_vpc_security_groups()}}
}
\concept{security groups}
