% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing.R
\name{billing}
\alias{billing}
\title{Fetch billing data}
\usage{
billing(date_start, date_end = as.character(Sys.Date()))
}
\arguments{
\item{date_start, date_end}{Start and end date to get billing data for.
Date format expected: \code{YYYY-MM-DD}}
}
\description{
Fetch billing data
}
\examples{
\dontrun{
billing(date_start="2023-01-01")
}
}
