% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bucket_arn}
\alias{bucket_arn}
\title{Get bucket ARN}
\usage{
bucket_arn(bucket, objects = "")
}
\arguments{
\item{bucket}{(character) a bucket name. required.}

\item{objects}{(character) path for object(s). default: \code{""}}
}
\value{
character string of bucket arn
}
\description{
Get bucket ARN
}
\examples{
bucket_arn("somebucket")
bucket_arn("somebucket", objects = "*")
bucket_arn("somebucket", objects = "data.csv")
bucket_arn("somebucket", objects = "myfolder/subset/data.csv")
bucket_arn("somebucket", objects = "myfolder/subset/*")
}
