% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{bucket_create_if_not}
\alias{bucket_create_if_not}
\title{Create a bucket if it does not exist}
\usage{
bucket_create_if_not(bucket, force = FALSE)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{force}{(logical) force deletion without going through the prompt.
default: \code{FALSE}. Should only be set to \code{TRUE} when required for
non-interactive use.}
}
\description{
Create a bucket if it does not exist
}
\keyword{internal}
