% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bucket_region}
\alias{bucket_region}
\title{Get bucket region}
\usage{
bucket_region(bucket)
}
\value{
character string of bucket region; NULL if bucket not found
}
\description{
Get bucket region
}
\keyword{internal}
