% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-redshift.R
\name{cluster_con_info}
\alias{cluster_con_info}
\title{Get connection information for all clusters}
\usage{
cluster_con_info(id)
}
\arguments{
\item{id}{(character) Cluster identifier. Use this identifier to refer to
the cluster for any subsequent cluster operations such as deleting or
modifying. The identifier also appears in the Amazon Redshift console.
Must be unique for all clusters within a Amazon Web Services account.}
}
\value{
a list of cluster details
}
\description{
Get connection information for all clusters
}
\keyword{internal}
