% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-redshift.R
\name{cluster_details}
\alias{cluster_details}
\title{Get information for all clusters}
\usage{
cluster_details()
}
\value{
a list of cluster details, see link below for format, with slots:
\itemize{
\item Marker (for pagination)
\item Clusters (each cluster; empty list if no clusters)
}
}
\description{
Get information for all clusters
}
\references{
\url{https://www.paws-r-sdk.com/docs/redshift_describe_clusters/}
}
\keyword{internal}
