% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{con_ec2}
\alias{con_ec2}
\title{Get the \code{paws} EC2 client}
\usage{
con_ec2()
}
\value{
a list with methods for interfacing with EC2;
see \url{https://www.paws-r-sdk.com/docs/ec2/}
}
\description{
Primarily for usage of VPC security groups
}
\seealso{
Other security groups: 
\code{\link{aws_vpc_sec_group_rules}()},
\code{\link{aws_vpc_security_group_create}()},
\code{\link{aws_vpc_security_group_ingress}()},
\code{\link{aws_vpc_security_groups}()},
\code{\link{aws_vpc_security_group}()},
\code{\link{aws_vpc_sg_with_ingresss}()}
}
\concept{security groups}
\keyword{internal}
