% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3con.R
\name{con_s3}
\alias{con_s3}
\title{Get the \code{paws} S3 client}
\usage{
con_s3()
}
\value{
a list with methods for interfacing with IAM;
see \url{https://www.paws-r-sdk.com/docs/s3/}
}
\description{
Get the \code{paws} S3 client
}
\keyword{internal}
