% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clients.R
\name{con_s3fs}
\alias{con_s3fs}
\title{s3fs connection}
\usage{
con_s3fs()
}
\value{
An S3 list with class 'sixtyfour_client'
}
\description{
s3fs connection
}
\details{
we set \code{refresh=TRUE} on \code{\link[s3fs:s3_file_system]{s3fs::s3_file_system()}} so that
you can change the s3 interface within an R session

You can toggle the interface set for one of minio, localstack, aws.
See \link{connections} for more information.
}
\examples{
\dontshow{if (aws_has_creds()) withAutoprint(\{ # examplesIf}
con <- con_s3fs()
con
con_s3fs()$file_copy
\dontshow{\}) # examplesIf}
}
\seealso{
\link{paws_clients}
}
