% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{con_sm}
\alias{con_sm}
\title{Get the \code{paws} Secrets Manager client}
\usage{
con_sm()
}
\value{
a list with methods for interfacing with Secrets Manager;
see \url{https://www.paws-r-sdk.com/docs/secretsmanager/}
}
\description{
Get the \code{paws} Secrets Manager client
}
\keyword{internal}
