% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{construct_db_secret}
\alias{construct_db_secret}
\title{Construct a database secret string or raw version of it}
\usage{
construct_db_secret(
  engine,
  host = "",
  username = "",
  password = "",
  dbname = "",
  port = "",
  as = "string"
)
}
\arguments{
\item{engine, host, username, password, dbname, port}{supply parameters to
go into either a json string or raw version of the json string}

\item{as}{(character) one of "string" or "raw"}
}
\description{
Construct a database secret string or raw version of it
}
\references{
\url{https://docs.aws.amazon.com/secretsmanager/latest/userguide/reference_secret_json_structure.html} # nolint
}
\keyword{internal}
