% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{equal_lengths}
\alias{equal_lengths}
\title{Check if two vectors have equal lengths}
\usage{
equal_lengths(x, y)
}
\description{
Check if two vectors have equal lengths
}
\details{
package \pkg{s3fs} is not checking that
\code{length(remote_path) == length(path)} as it purports to be doing -
thus the need for this helper function to check this
}
\keyword{internal}
