% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{figure_out_policy_arn}
\alias{figure_out_policy_arn}
\title{Figure out policy Arn from a name}
\usage{
figure_out_policy_arn(name)
}
\arguments{
\item{name}{(character) a policy name. required.}
}
\value{
\code{NULL} when not found; otherwise an ARN string
}
\description{
Figure out policy Arn from a name
}
\examples{
\dontshow{if (aws_has_creds()) withAutoprint(\{ # examplesIf}
# aws managed
figure_out_policy_arn("AmazonS3ReadOnlyAccess")
# aws managed, job function
figure_out_policy_arn("Billing")
figure_out_policy_arn("DataScientist")
# doesn't exist
figure_out_policy_arn("DoesNotExist")
\dontshow{\}) # examplesIf}
}
