% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first}
\alias{first}
\title{Get the first element of a vector}
\usage{
first(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the first element of the vector
}
\description{
Get the first element of a vector
}
\keyword{internal}
