% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{group_policies}
\alias{group_policies}
\title{Preset group policies}
\usage{
group_policies(group)
}
\arguments{
\item{group}{(character)}
}
\value{
character vector of policy names
}
\description{
Preset group policies
}
\section{Admin group policies}{

\itemize{
\item AdministratorAccess
\item Billing
\item CostOptimizationHubAdminAccess
\item AWSBillingReadOnlyAccess
\item AWSCostAndUsageReportAutomationPolicy
}
}

\section{User group policies}{

\itemize{
\item AmazonRDSReadOnlyAccess
\item AmazonRedshiftReadOnlyAccess
\item AmazonS3ReadOnlyAccess
\item AWSBillingReadOnlyAccess
\item IAMReadOnlyAccess
}
}

\examples{
group_policies("admin")
group_policies("users")
}
