% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{has_access}
\alias{has_access}
\title{Check if a user has access to an AWS service}
\usage{
has_access(fun, ...)
}
\arguments{
\item{fun}{(funcction) a function. required}

\item{...}{additional named args passed to \code{fun}}
}
\value{
single boolean. checks \code{\link[rlang:type-predicates]{rlang::is_null()}} against \verb{$error} result of
call to \code{\link[purrr:safely]{purrr::safely()}}
}
\description{
Check if a user has access to an AWS service
}
\details{
really just a generic check that any function can run with
its inputs; not specific to AWS or any particular function
}
