% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-rds.R
\name{instance_details}
\alias{instance_details}
\title{Get information for all RDS instances}
\usage{
instance_details()
}
\value{
a list of instance details
}
\description{
Get information for all RDS instances
}
\keyword{internal}
