% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{ip_address}
\alias{ip_address}
\title{Get your IP address using \url{https://ifconfig.me/ip}}
\usage{
ip_address()
}
\value{
(character) ip address
}
\description{
Get your IP address using \url{https://ifconfig.me/ip}
}
\keyword{internal}
