% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{ip_permissions_generator}
\alias{ip_permissions_generator}
\title{Ip Permissions generator}
\usage{
ip_permissions_generator(engine, port = NULL, description = NULL)
}
\arguments{
\item{engine}{(character) one of mariadb, mysql, or postgres}

\item{port}{(character) port number. port determined from \code{engine}
if \code{port} not given. default: \code{NULL}}

\item{description}{(character) description. if not given, autogenerated
depending on value of \code{engine}}
}
\value{
a list with slots: FromPort, ToPort, IpProtocol, and IpRanges
}
\description{
Ip Permissions generator
}
