% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last}
\alias{last}
\title{Get the last element of a vector}
\usage{
last(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the last element of the vector
}
\description{
Get the last element of a vector
}
\keyword{internal}
