% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paginate_aws}
\alias{paginate_aws}
\title{Paginate over list_* methods}
\usage{
paginate_aws(fun, target, ...)
}
\arguments{
\item{fun}{(function) a function to call}

\item{target}{(character) a list element to get}

\item{...}{named args passed on to \code{fun}}
}
\description{
Paginate over list_* methods
}
\examples{
\dontrun{
# FIXME: could remove target param and poach the name of the fun
# e.g,. from list_roles we can get Roles
paginate_aws(fun = env64$iam$list_roles, target = "Roles")
paginate_aws(fun = env64$iam$list_policies, target = "Policies")
}
}
\keyword{internal}
