% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paginate_aws_token}
\alias{paginate_aws_token}
\title{Paginate over list_* methods with NextToken}
\usage{
paginate_aws_token(fun, target, ...)
}
\arguments{
\item{fun}{(character) the name of a function to call - not the function
itself}

\item{target}{(character) a list element to get}

\item{...}{named args passed on to \code{fun}}
}
\description{
Paginate over list_* methods with NextToken
}
\note{
\code{paginate_aws_token(fun=con_sm()$list_secrets, target="SecretList")}
}
\keyword{internal}
