% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{path_from}
\alias{path_from}
\title{Get file path starting at a certain path component}
\usage{
path_from(path, from)
}
\description{
Get file path starting at a certain path component
}
\examples{
path_from(path = "Rtmpxsqth0/apples/mcintosh/orange.csv", from = "apples")
}
\keyword{internal}
