% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_s3_build}
\alias{path_s3_build}
\title{Build s3 paths}
\usage{
path_s3_build(x)
}
\arguments{
\item{x}{unnamed list of parsed paths, from \code{\link[=path_s3_parse]{path_s3_parse()}}}
}
\description{
Build s3 paths
}
\examples{
paths <- c(
  "s3://s64-test-2/DESCRIPTION",
  "s3://s64-test-2/stuff.txt",
  "s3://s64-test-2/some/other/path/things.csv"
)
x <- path_s3_parse(paths)
path_s3_build(x)
}
\keyword{internal}
