% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_s3_parse}
\alias{path_s3_parse}
\title{Parse s3 paths}
\usage{
path_s3_parse(paths)
}
\arguments{
\item{paths}{(character) one or more s3 paths}
}
\value{
an unnamed list with each slot a named list with bucket, path,
and file
}
\description{
Parse s3 paths
}
\examples{
path_s3_parse("s3://s64-test-2/DESCRIPTION")
path_s3_parse("s3://s64-test-2/some/other/path/things.csv")
paths <- c(
  "s3://s64-test-2/DESCRIPTION",
  "s3://s64-test-2/stuff.txt",
  "s3://s64-test-2/some/other/path/things.csv"
)
path_s3_parse(paths)

# if a path is not an s3 path
paths <- c(
  "s3://s64-test-2/DESCRIPTION",
  "s3://s64-test-2/stuff.txt",
  "s64-test-2/some/other/path/things.csv"
)
try(path_s3_parse(paths))
}
\keyword{internal}
