% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random_bucket}
\alias{random_bucket}
\title{Get a random bucket name}
\usage{
random_bucket(prefix = "bucket-", length = 16)
}
\arguments{
\item{prefix}{(character) any string. required.}
}
\value{
(character) a bucket name prefixed with \code{prefix}
(default: "bucket-")
}
\description{
Get a random bucket name
}
\examples{
random_bucket()
replicate(10, random_bucket())
}
