% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random_user}
\alias{random_user}
\alias{random_role}
\title{Get a random user or role}
\usage{
random_user()

random_role()
}
\value{
(character) a username or role name with a random adjective plus a
random noun combined into one string, shortened to no longer than 16
characters if longer than 16
}
\description{
Get a random user or role
}
\examples{
random_user()
random_role()
replicate(10, random_user())
}
