% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_actions_full}
\alias{s3_actions_full}
\title{S3 actions for full access (read and write), from the AWS
managed policy \code{AmazonS3FullAccess}}
\usage{
s3_actions_full()
}
\value{
character vector of actions
}
\description{
S3 actions for full access (read and write), from the AWS
managed policy \code{AmazonS3FullAccess}
}
\keyword{internal}
