% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_actions_read}
\alias{s3_actions_read}
\title{S3 actions for reading, from the AWS managed policy
\code{AmazonS3ReadOnlyAccess}}
\usage{
s3_actions_read()
}
\value{
character vector of actions
}
\description{
S3 actions for reading, from the AWS managed policy
\code{AmazonS3ReadOnlyAccess}
}
\keyword{internal}
