% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{s3fs_creds_refresh}
\alias{s3fs_creds_refresh}
\title{Refresh creds for the s3fs package}
\usage{
s3fs_creds_refresh()
}
\value{
nothing, updates creds with \code{\link[s3fs:s3_file_system]{s3fs::s3_file_system()}}
}
\description{
Refresh creds for the s3fs package
}
\details{
utility function to update creds for use with any
\code{s3fs} functions. We do load creds for \code{s3fs} on package load
but if creds are changed mid-R session, then we would still be
using the creds used at package load time
}
\keyword{internal}
