% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixtyfour-package.R
\docType{data}
\name{service_map}
\alias{service_map}
\title{Mapping of full names of AWS services to acronyms}
\format{
\subsection{\code{service_map}}{

A data frame with 178 rows and 2 columns:
\describe{
\item{service}{Service name in full}
\item{acronym}{The acronym, from 2 to 5 characters in length}
...
}
}
}
\source{
\url{https://tommymaynard.com/aws-service-acronyms/}
}
\usage{
service_map
}
\description{
Mapping of full names of AWS services to acronyms
}
\keyword{datasets}
