% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_clients.R
\name{set_iam_client}
\alias{set_iam_client}
\title{Set client}
\usage{
set_iam_client(testing = FALSE)
}
\arguments{
\item{testing}{(logical) whether to set iam client for testing.
default: \code{FALSE}}
}
\description{
Set client
}
\keyword{internal}
