% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{set_s3_interface}
\alias{set_s3_interface}
\title{Set the S3 compatible interface to use}
\usage{
set_s3_interface(interface = "aws")
}
\arguments{
\item{interface}{the s3 compatible interface to use.
options: "aws" (default), "minio", or "localstack"}
}
\value{
a \code{paws} s3 client object of class \code{list}
}
\description{
Set the S3 compatible interface to use
}
\details{
This function sets the connection details for both \code{paws} and \code{s3fs}
packages, but only returns something for \code{paws}

This function expects to have access to the following environment
variables:
\itemize{
\item \code{AWS_ACCESS_KEY_ID}
\item \code{AWS_SECRET_ACCESS_KEY}
\item \code{AWS_REGION}
\item \code{MINIO_USER}
\item \code{MINIO_PWD}
\item \code{MINIO_ENDPOINT}
}

For \code{\link[s3fs:s3_file_system]{s3fs::s3_file_system()}} we set \code{refresh=TRUE} so that
you can change the s3 interface within an R session.
}
\keyword{internal}
