% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{six_group_delete}
\alias{six_group_delete}
\title{Delete a group, magically}
\usage{
six_group_delete(name)
}
\arguments{
\item{name}{(character) A group name. required}
}
\value{
\code{NULL} invisibly
}
\description{
Delete a group, magically
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_group/}
docs for more details
}
\examples{
\dontshow{if (aws_has_creds()) withAutoprint(\{ # examplesIf}
group <- random_string("group")
aws_group_create(group)
six_group_delete(group)
\dontshow{\}) # examplesIf}
}
\seealso{
Other groups:
\code{\link[=aws_group]{aws_group()}},
\code{\link[=aws_group_create]{aws_group_create()}},
\code{\link[=aws_group_delete]{aws_group_delete()}},
\code{\link[=aws_group_exists]{aws_group_exists()}},
\code{\link[=aws_groups]{aws_groups()}}
}
\concept{groups}
