% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixtyfour-package.R
\docType{package}
\name{sixtyfour-package}
\alias{sixtyfour}
\alias{sixtyfour-package}
\title{sixtyfour: Humane Interface to Amazon Web Services}
\description{
An opinionated interface to Amazon Web Services \url{https://aws.amazon.com}, with functions for interacting with 'IAM' (Identity and Access Management), 'S3' (Simple Storage Service), 'RDS' (Relational Data Service), Redshift, and Billing. Lower level functions ('aws_' prefix) are for do it yourself workflows, while higher level functions ('six_' prefix) automate common tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/getwilds/sixtyfour}
  \item \url{https://getwilds.org/sixtyfour/}
  \item Report bugs at \url{https://github.com/getwilds/sixtyfour/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{sachamber@fredhutch.org} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})

Authors:
\itemize{
  \item Scott Chamberlain \email{sachamber@fredhutch.org} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})
  \item Sean Kross
}

Other contributors:
\itemize{
  \item Fred Hutchinson Cancer Center \email{wilds@fredhutch.org} (\href{https://ror.org/007ps6h72}{ROR}) [funder]
}

}
\keyword{internal}
