% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fetch_secret.R
\name{ui_fetch_secret}
\alias{ui_fetch_secret}
\title{Fetch secrets}
\usage{
ui_fetch_secret(user = NULL, password = NULL, engine = NULL)
}
\arguments{
\item{user}{(character) xx}

\item{password}{(character) password}

\item{engine}{(character) engine to filter secrets with. if not supplied
(\code{NULL}) all secrets are considered}
}
\description{
Fetch secrets
}
\section{How the function works}{

\itemize{
\item If user and password are supplied they are returned immediately
\item If user and password are not supplied, we fetch all secrets in
your secrets manager service, and then ask you which one you'd like
to use. If you choose none of them this function returns NULL for
both user and password
}
}

\examples{
\dontrun{
# user,pwd supplied, return them right away at top of fxn
ui_fetch_secret(engine = "mariadb", user = "jane", password = "apple")

# user,pwd null
ui_fetch_secret(engine = "redshift")
ui_fetch_secret(engine = "mariadb")
}
}
\keyword{internal}
