% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{user_list_cleanup}
\alias{user_list_cleanup}
\title{User list cleanup - internal function}
\usage{
user_list_cleanup(x)
}
\arguments{
\item{x}{what is this param Sean? some kind of tibble presumably}
}
\description{
User list cleanup - internal function
}
