% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{wait_for_cluster}
\alias{wait_for_cluster}
\title{Wait for a Redshift cluster to have a certain status}
\usage{
wait_for_cluster(id, sleep = 2, status_target = "available")
}
\arguments{
\item{id}{(character) Cluster identifier. Use this identifier to refer to
the cluster for any subsequent cluster operations such as deleting or
modifying. The identifier also appears in the Amazon Redshift console.
Must be unique for all clusters within a Amazon Web Services account.}

\item{sleep}{(integer/numeric) number of seconds to wait between
checks of the cluster status (i.e., http requests)}

\item{status_target}{(character) status to wait for. default: "available"}

\item{fun}{(function) a function to check status of something;
must return a single boolean, e.g., \code{aws_db_cluster_status} or
\code{aws_db_instance_status}}
}
\value{
nothing, exits if there's an error, or if the while
loop completes
}
\description{
Wait for a Redshift cluster to have a certain status
}
\examples{
\dontrun{
wait_for_cluster(id = "scotts-test-cluster-456")
}
}
\keyword{internal}
