% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{wait_until}
\alias{wait_until}
\title{wait fxn generator}
\usage{
wait_until(fun, message)
}
\arguments{
\item{fun}{(function) a function to check status of something;
must return a single boolean, e.g., \code{aws_db_cluster_status} or
\code{aws_db_instance_status}}

\item{message}{(character) the message to print at the beginning
of \code{cli::cli_progress_bar}}
}
\description{
wait fxn generator
}
\keyword{internal}
