% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{wait_until_cluster_available}
\alias{wait_until_cluster_available}
\title{Wait until a cluster has status "available"}
\usage{
wait_until_cluster_available(id, sleep = 3)
}
\arguments{
\item{id}{(characteR) Cluster identifier. Use this identifier to refer to
the cluster for any subsequent cluster operations such as deleting or
modifying. The identifier also appears in the Amazon Redshift console.
Must be unique for all clusters within a Amazon Web Services account.}

\item{sleep}{(integer/numeric) number of seconds to wait between
checks of the cluster status (i.e., http requests)}
}
\value{
nothing, exits if there's an error, or if the while
loop completes
}
\description{
Wait until a cluster has status "available"
}
\examples{
\dontrun{
wait_until_cluster_available(id = "scotts-test-cluster-456")
}
}
\keyword{internal}
