% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{without_verbose}
\alias{without_verbose}
\title{Without verbose output}
\usage{
without_verbose(code)
}
\arguments{
\item{code}{(expression) Code to run without verbose output.}
}
\description{
Without verbose output
}
