% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{account_id}
\alias{account_id}
\title{Get account ID of current user}
\usage{
account_id()
}
\value{
list with 3 elements:
\itemize{
\item UserId: the ID for the user
\item Account: account ID the user is in
\item Arn: arn for the user
}
}
\description{
Get account ID of current user
}
\keyword{internal}
