% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{as_policy_arn}
\alias{as_policy_arn}
\title{Convert a policy name to a policy ARN}
\usage{
as_policy_arn(name)
}
\arguments{
\item{name}{(character) a policy name or arn}
}
\value{
a policy ARN
}
\description{
Convert a policy name to a policy ARN
}
\note{
uses exact matching; fails with error if there's no match;
beware as there is no validation is done of a user input policy arn
}
\examples{
\dontrun{
as_policy_arn("ReadOnlyAccess")
as_policy_arn("arn:aws:iam::aws:policy/ReadOnlyAccess")
as_policy_arn("AmazonRDSDataFullAccess")
# as_policy_arn("Blarp")
# as_policy_arn(letters)
# as_policy_arn(5)
}
}
\seealso{
Other policies: 
\code{\link{aws_policies}()},
\code{\link{aws_policy_attach}()},
\code{\link{aws_policy_create}()},
\code{\link{aws_policy_detach}()},
\code{\link{aws_policy_exists}()},
\code{\link{aws_policy}()}
}
\concept{policies}
