% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_acl_modify}
\alias{aws_bucket_acl_modify}
\title{Modify a bucket ACL}
\usage{
aws_bucket_acl_modify(bucket, ...)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{...}{named params passed on to
\url{https://www.paws-r-sdk.com/docs/s3_put_bucket_acl/}}
}
\description{
Modify a bucket ACL
}
\examples{
\dontrun{
aws_bucket_acl_modify("s3://s64-test-2")
}
}
