% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{aws_bucket_change_user}
\alias{aws_bucket_change_user}
\title{Change user or group permissions for a bucket}
\usage{
aws_bucket_change_user(bucket, username, permissions)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{username}{(character) A user name. required}

\item{permissions}{(character) user permissions, one of
read or write. write includes read}
}
\value{
xxx
}
\description{
Change user or group permissions for a bucket
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_bucket_change_user("mybucket", "sean", "write")
aws_bucket_change_user("mybucket", "sean", c("write", "read"))
\dontshow{\}) # examplesIf}
}
