% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_create}
\alias{aws_bucket_create}
\title{Create an S3 bucket}
\usage{
aws_bucket_create(bucket, ...)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_create_bucket/}{list_objects}}
}
\description{
Create an S3 bucket
}
\examples{
\dontrun{
aws_bucket_create(bucket="s64-test-2")
}
}
