% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_delete}
\alias{aws_bucket_delete}
\title{Delete an S3 bucket}
\usage{
aws_bucket_delete(bucket, ...)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_delete_bucket/}{delete_bucket}}
}
\value{
an empty list
}
\description{
Delete an S3 bucket
}
\note{
Requires the env var \code{AWS_REGION}. This function prompts you to make
sure that you want to delete the bucket.
}
\examples{
\dontrun{
aws_bucket_create(bucket="bucket-to-delete-111")
aws_buckets()
aws_bucket_delete(bucket="bucket-to-delete-111")
aws_buckets()
}
}
