% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_download}
\alias{aws_bucket_download}
\title{Download an S3 bucket}
\usage{
aws_bucket_download(bucket, dest_path, ...)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{dest_path}{(character) destination directory to store files. required}

\item{...}{named parameters passed on to \code{\link[s3fs:download]{s3fs::s3_dir_download()}}}
}
\description{
Download an S3 bucket
}
\note{
Requires the env var \code{AWS_REGION}. This function prompts you to make
sure that you want to delete the bucket.
}
\examples{
\dontrun{
aws_bucket_create(bucket = "tmp-bucket-369")
desc_file <- file.path(system.file(), "DESCRIPTION")
aws_file_upload(bucket = "tmp-bucket-369", path = desc_file)
aws_file_upload(bucket = "tmp-bucket-369", path = desc_file, key = "d_file")
temp_dir <- file.path(tempdir(), "tmp-bucket-369")
aws_bucket_download(bucket = "tmp-bucket-369", dest_path = temp_dir)

# cleanup
aws_bucket_delete("tmp-bucket-369")
}
}
\seealso{
Other buckets: 
\code{\link{aws_bucket_create}()},
\code{\link{aws_bucket_delete}()},
\code{\link{aws_bucket_exists}()},
\code{\link{aws_bucket_list_objects}()},
\code{\link{aws_bucket_tree}()},
\code{\link{aws_bucket_upload}()},
\code{\link{aws_buckets}()}
}
\concept{buckets}
