% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_list_objects}
\alias{aws_bucket_list_objects}
\title{List objects in an S3 bucket}
\usage{
aws_bucket_list_objects(bucket, ...)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_list_objects/}{list_objects}}
}
\description{
List objects in an S3 bucket
}
\examples{
\dontrun{
aws_bucket_list_objects(bucket="s64-test-2")
}
}
