% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_tree}
\alias{aws_bucket_tree}
\title{Print a tree of the objects in a bucket}
\usage{
aws_bucket_tree(bucket, recurse = TRUE, ...)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{recurse}{(logical) returns all AWS S3 objects in lower sub
directories, default: \code{TRUE}}

\item{...}{Additional arguments passed to \code{\link[s3fs:s3_dir_tree]{s3fs::s3_dir_tree()}}}
}
\value{
character vector of objects/files within the bucket,
printed as a tree
}
\description{
Print a tree of the objects in a bucket
}
\examples{
\dontrun{
aws_bucket_tree("s3://s64-test-2")
}
}
