% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_bucket_upload}
\alias{aws_bucket_upload}
\title{Upload a folder of files to create an S3 bucket}
\usage{
aws_bucket_upload(path, bucket, max_batch = fs::fs_bytes("100MB"), ...)
}
\arguments{
\item{path}{(character) local path to a directory. required}

\item{bucket}{(character) bucket name. required}

\item{max_batch}{(fs_bytes) maximum batch size being uploaded with each
multipart}

\item{...}{named parameters passed on to \code{\link[s3fs:upload]{s3fs::s3_dir_upload()}}}
}
\description{
Upload a folder of files to create an S3 bucket
}
\note{
Requires the env var \code{AWS_REGION}. This function prompts you to make
sure that you want to delete the bucket.
}
\examples{
\dontrun{
library(fs)
tdir <- path(tempdir(), "apples")
dir.create(tdir)
tfiles <- replicate(n=10, file_temp(tmp_dir = tdir, ext=".txt"))
invisible(lapply(tfiles, function(x) write.csv(mtcars, x)))

aws_bucket_upload(path=tdir, bucket="a-new-bucket-345")
aws_bucket_list_objects("a-new-bucket-345")

# cleanup
objs <- aws_bucket_list_objects("a-new-bucket-345")
aws_file_delete(objs$uri)
aws_bucket_delete("a-new-bucket-345")
}
}
