% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_buckets}
\alias{aws_buckets}
\title{List S3 buckets}
\usage{
aws_buckets(...)
}
\arguments{
\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_list_buckets/}{list_buckets}}
}
\value{
tibble with zero or more rows (each an S3 bucket), with two columns:
\itemize{
\item Name (character)
\item CreationDate (dttm)
}
}
\description{
List S3 buckets
}
\examples{
\dontrun{
aws_buckets()
}
}
