% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{aws_buckets}
\alias{aws_buckets}
\title{List S3 buckets}
\usage{
aws_buckets(...)
}
\arguments{
\item{...}{named parameters passed on to \code{\link[s3fs:info]{s3fs::s3_dir_info()}}}
}
\value{
if no objects found, an empty tibble. if tibble has rows each
is an S3 bucket, with 8 columns:
\itemize{
\item bucket_name (character)
\item key (character)
\item uri (character)
\item size (fs::bytes)
\item type (character)
\item owner (character)
\item etag (character)
\item last_modified (dttm)
}
}
\description{
List S3 buckets
}
\note{
we set \code{refresh=TRUE} internally to make sure we return up to date
information about your buckets rather than what's cached locally
}
\examples{
\dontrun{
aws_buckets()
}
}
