% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-rds.R
\name{aws_db_rds_client}
\alias{aws_db_rds_client}
\title{Get the \code{paws} RDS client}
\usage{
aws_db_rds_client()
}
\value{
a list with methods for interfacing with RDS;
see \url{https://www.paws-r-sdk.com/docs/rds/}
}
\description{
Get the \code{paws} RDS client
}
\note{
returns existing client if found; a new client otherwise
}
\seealso{
Other database: 
\code{\link{aws_db_cluster_status}()},
\code{\link{aws_db_instance_status}()},
\code{\link{aws_db_rds_con}()},
\code{\link{aws_db_rds_create}()},
\code{\link{aws_db_redshift_client}()},
\code{\link{aws_db_redshift_con}()},
\code{\link{aws_db_redshift_create}()}
}
\concept{database}
