% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{aws_db_rds_con}
\alias{aws_db_rds_con}
\title{Get a database connection to Amazon RDS}
\usage{
aws_db_rds_con(...)
}
\arguments{
\item{...}{named parameters passed on to
\href{https://dbi.r-dbi.org/reference/dbconnect}{DBI::dbConnect}}
}
\description{
Get a database connection to Amazon RDS
}
\note{
RDS supports: Aurora (both PostgreSQL and MySQL compatible),
PostgreSQL, MariaDB, MySQL, Orace, MS SQL Server
}
