% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{aws_db_redshift_con}
\alias{aws_db_redshift_con}
\title{Get a database connection to Amazon Redshift}
\usage{
aws_db_redshift_con(host, port, dbname, user, pwd, ...)
}
\arguments{
\item{host, port, dbname, user, pwd, ...}{named parameters passed on to
\href{https://dbi.r-dbi.org/reference/dbconnect}{DBI::dbConnect}}
}
\description{
Get a database connection to Amazon Redshift
}
\details{
The connection returned is created using
\href{https://rpostgres.r-dbi.org/}{RPostgres}

You can manage Redshift programatically via
\href{https://www.paws-r-sdk.com/docs/redshift/}{paws::redshift}
}
\examples{
\dontrun{
library(DBI)
library(RPostgres)

con <- aws_db_redshift_con()
con

con <- aws_db_rds_con()
con
}
}
