% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{aws_db_redshift_create}
\alias{aws_db_redshift_create}
\title{Create a Redshift cluster}
\usage{
aws_db_redshift_create(
  id,
  user,
  pwd,
  db_name = "dev",
  cluster_type = "multi-node",
  node_type = "dc2.large",
  number_nodes = 2,
  ...
)
}
\arguments{
\item{id}{(character) Cluster identifier. Use this identifier to refer to the
cluster for any subsequent cluster operations such as deleting or modifying.
The identifier also appears in the Amazon Redshift console. Must be unique
for all clusters within an Amazon Web Services account.}

\item{user}{(character) User name associated with the admin user account for
the cluster that is being created.}

\item{pwd}{(character) Password associated with the admin user account for
the cluster that is being created.}

\item{db_name}{(character) The name of the first database to be created when
the cluster is created. default: "dev"}

\item{cluster_type}{(character) The type of the cluster: "single-node" or
"multi-node" (default).}

\item{node_type}{(character) The node type to be provisioned for the cluster.
defaul: "dc2.large"}

\item{number_nodes}{(integer/numeric) number of nodes; for multi-node
cluster type, this must be 2 or greater. default: 2}

\item{...}{named parameters passed on to
\href{https://www.paws-r-sdk.com/docs/redshift_create_cluster/}{create_cluster}}
}
\value{
a redshift class from \code{paws} with methods on the object;
see \url{https://www.paws-r-sdk.com/docs/redshift/}
}
\description{
!NOT READY YET!
}
\note{
See above link to \code{create_cluster} docs for details on requirements
for each parameter
}
