% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_attr}
\alias{aws_file_attr}
\title{File attributes}
\usage{
aws_file_attr(bucket, key, ...)
}
\arguments{
\item{bucket}{(character) an S3 bucket. required}

\item{key}{(character) a key for an object in an S3 \code{bucket}. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_head_object/}{head_object}}
}
\value{
\code{list} of length 0
}
\description{
File attributes
}
\examples{
\dontrun{
aws_file_attr(bucket = "s64-test-2", key = "DESCRIPTION")
aws_file_attr(bucket = "s64-test-2", key = "ddd")
aws_file_attr(bucket = "s64-test-2", key = "doesntexist")
}
}
