% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_attr}
\alias{aws_file_attr}
\title{File attributes}
\usage{
aws_file_attr(remote_path)
}
\arguments{
\item{remote_path}{(character) one or more remote S3 paths. required}
}
\value{
a tibble with many columns, with number of rows matching length
of \code{remote_path}
}
\description{
File attributes
}
\note{
uses \code{\link[s3fs:info]{s3fs::s3_file_info()}} internally
}
\examples{
\dontrun{
# files one by one
aws_file_attr(s3_path("s64-test-2", "DESCRIPTION"))
aws_file_attr(s3_path("s64-test-2", "ddd"))
aws_file_attr(s3_path("s64-test-2", "doesntexist"))
# or all together
aws_file_attr(s3_path("s64-test-2", c("DESCRIPTION", "ddd")))
}
}
\seealso{
Other files: 
\code{\link{aws_file_copy}()},
\code{\link{aws_file_delete}()},
\code{\link{aws_file_download}()},
\code{\link{aws_file_exists}()},
\code{\link{aws_file_rename}()},
\code{\link{aws_file_upload}()}
}
\concept{files}
