% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_delete}
\alias{aws_file_delete}
\title{Delete a file}
\usage{
aws_file_delete(remote_path, ...)
}
\arguments{
\item{remote_path}{(character) one or more remote S3 paths. required}

\item{...}{named parameters passed on to
\href{https://www.paws-r-sdk.com/docs/s3_delete_object/}{delete_object}}
}
\value{
\code{NULL} invisibly
}
\description{
Delete a file
}
\examples{
\dontrun{
# create a file
tfile <- tempfile()
cat("Hello World!", file = tfile)
aws_file_upload(remote_path = "s3://s64-test-2", path = tfile)

# delete the file
aws_file_delete(s3_path("s64-test-2", basename(tfile)))

# file does not exist - no error is raised
aws_file_delete(s3_path("s64-test-2", "TESTING123"))
}
}
\seealso{
Other files: 
\code{\link{aws_file_attr}()},
\code{\link{aws_file_copy}()},
\code{\link{aws_file_download}()},
\code{\link{aws_file_exists}()},
\code{\link{aws_file_rename}()},
\code{\link{aws_file_upload}()},
\code{\link{six_file_upload}()}
}
\concept{files}
