% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_download}
\alias{aws_file_download}
\title{Download a file}
\usage{
aws_file_download(bucket, key, path, ...)
}
\arguments{
\item{bucket}{(character) an S3 bucket. required}

\item{key}{(character) a key for an object in an S3 \code{bucket}. required}

\item{path}{(character) a file path to read from or write to. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_download_file/}{download_file}}
}
\value{
\code{list} of length 0
}
\description{
Download a file
}
\details{
Wraps \href{https://www.paws-r-sdk.com/docs/s3_download_file/}{download_file}
}
\examples{
\dontrun{
temp_path <- tempfile()
aws_file_download(bucket = "s64-test-2", key = "DESCRIPTION",
path = temp_path)

# S3 key doesn't exist
aws_file_download(bucket = "s64-test-2", key = "TESTING123",
path = temp_path)
}
}
