% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_download}
\alias{aws_file_download}
\title{Download a file}
\usage{
aws_file_download(remote_path, path, ...)
}
\arguments{
\item{remote_path}{(character) one or more remote S3 paths. required}

\item{path}{(character) one or more file paths to write to. required}

\item{...}{named parameters passed on to \code{\link[s3fs:download]{s3fs::s3_file_download()}}}
}
\value{
(character) a vector of local file paths
}
\description{
Download a file
}
\note{
USES A FORK OF s3fs FOR A MINOR FIX THAT MAKES LENGTH>1 INPUTS WORK
}
\examples{
\dontrun{
tfile <- tempfile()
aws_file_download(remote_path = "s3://s64-test-2/DESCRIPTION", path = tfile)

# many files
tfiles <- replicate(n = 3, tempfile())
aws_file_download(remote_path =
 s3_path("s64-test-2", c("a_file", "c_file", "d_file")), path = tfiles)

## length of `remote_path` and `path` must be the same
tfiles <- replicate(n = 2, tempfile())
aws_file_download(remote_path =
 s3_path("s64-test-2", c("a_file", "c_file", "d_file")), path = tfiles)

# S3 file does not exist
temp_path <- tempfile()
aws_file_download(s3_path("s64-test-2", "TESTING123"), temp_path)
}
}
