% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_exists}
\alias{aws_file_exists}
\title{Check if a file exists}
\usage{
aws_file_exists(bucket, key, ...)
}
\arguments{
\item{bucket}{(character) an S3 bucket. required}

\item{key}{(character) a key for an object in an S3 \code{bucket}. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_put_object/}{put_object}}
}
\value{
TRUE or FALSE
}
\description{
Check if a file exists
}
\examples{
\dontrun{
aws_file_exists(bucket = "s64-test-2", key = "DESCRIPTION")
aws_file_exists(bucket = "s64-test-2", key = "doesntexist")
}
}
