% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_exists}
\alias{aws_file_exists}
\title{Check if a file exists}
\usage{
aws_file_exists(remote_path)
}
\arguments{
\item{remote_path}{(character) one or more remote S3 paths. required}
}
\value{
vector of booleans (\code{TRUE} or \code{FALSE}), length matches
\code{length(remote_path)}
}
\description{
Check if a file exists
}
\examples{
\dontrun{
aws_file_exists(s3_path("s64-test-2", "DESCRIPTION"))
aws_file_exists(s3_path("s64-test-2", "doesntexist"))
aws_file_exists(s3_path("s64-test-2", c("DESCRIPTION", "doesntexist")))
}
}
\seealso{
Other files: 
\code{\link{aws_file_attr}()},
\code{\link{aws_file_copy}()},
\code{\link{aws_file_delete}()},
\code{\link{aws_file_download}()},
\code{\link{aws_file_rename}()},
\code{\link{aws_file_upload}()}
}
\concept{files}
