% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_rename}
\alias{aws_file_rename}
\title{Rename a remote file}
\usage{
aws_file_rename(remote_path, new_remote_path, ...)
}
\arguments{
\item{remote_path}{(character) one or more remote S3 paths. required}

\item{new_remote_path}{(character) one or more remote S3 paths. required.
length must match \code{remote_path}}

\item{...}{named parameters passed on to \code{\link[s3fs:s3_file_move]{s3fs::s3_file_move()}}}
}
\value{
vector of paths, length matches \code{length(remote_path)}
}
\description{
Rename a remote file
}
\examples{
\dontrun{
aws_file_rename(s3_path("s64-test-2", "DESCRIPTION"),
  s3_path("s64-test-2", "DESC"))

tfiles <- replicate(n = 3, tempfile())
for (i in tfiles) cat("Hello\nWorld\n", file = i)
paths <- s3_path("s64-test-2", c("aaa", "bbb", "ccc"), ext = "txt")
aws_file_upload(tfiles, paths)
new_paths <- s3_path("s64-test-2", c("new_aaa", "new_bbb", "new_ccc"),
  ext = "txt")
aws_file_rename(paths, new_paths)
}
}
