% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_upload}
\alias{aws_file_upload}
\title{Upload a file}
\usage{
aws_file_upload(bucket, path, key = basename(path), ...)
}
\arguments{
\item{bucket}{(character) an S3 bucket. required}

\item{path}{(character) a file path to read from or write to. required}

\item{key}{(character) a key for an object in an S3 \code{bucket}. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/s3_put_object/}{put_object}}
}
\value{
a tibble with two columns and many rows
}
\description{
Upload a file
}
\details{
Wraps \href{https://www.paws-r-sdk.com/docs/s3_put_object/}{put_object}

\code{bucket} parameter:
\itemize{
\item For upload: if it does exist it will be created
\item For download: if it does not exist, function will return an error
}
}
\examples{
\dontrun{
desc_file <- file.path(system.file(), "DESCRIPTION")
aws_file_upload(bucket = "s64-test-2", path = desc_file)

# supply a different key
aws_file_upload(bucket = "s64-test-2", path = desc_file, key = "d_file")

# set expiration, expire 1 minute from now
aws_file_upload(bucket = "s64-test-2", path = desc_file, key = "ddd",
Expires = Sys.time() + 60)

# bucket doesn't exist
aws_file_upload(bucket = "not-a-bucket", path = desc_file)
# path doesn't exist
aws_file_upload(bucket = "s64-test-2", path = "file_doesnt_exist.txt")
}
}
