% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{aws_file_upload}
\alias{aws_file_upload}
\title{Upload a file}
\usage{
aws_file_upload(path, remote_path, ...)
}
\arguments{
\item{path}{(character) a file path to read from. required}

\item{remote_path}{(character) a remote path where the file
should go. required}

\item{...}{named parameters passed on to \code{\link[s3fs:copy]{s3fs::s3_file_copy()}}}
}
\value{
(character) a vector of remote s3 paths
}
\description{
Upload a file
}
\details{
\itemize{
\item For upload: if it does exist it will be created
\item For download: if it does not exist, function will return an error
}
}
\examples{
\dontrun{
demo_rds_file <- file.path(system.file(), "Meta/demo.rds")
aws_file_upload(
  demo_rds_file,
  s3_path("s64-test-2", basename(demo_rds_file))
)

## many files at once
links_file <- file.path(system.file(), "Meta/links.rds")
aws_file_upload(
  c(demo_rds_file, links_file),
  s3_path("s64-test-2", c(basename(demo_rds_file), basename(links_file)))
)

# set expiration, expire 1 minute from now
aws_file_upload(demo_rds_file, s3_path("s64-test-2", "ddd.rds"),
  Expires = Sys.time() + 60
)

# bucket doesn't exist
aws_file_upload(demo_rds_file, "s3://not-a-bucket/eee.rds")
}

\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# path doesn't exist
aws_file_upload(
  "file_doesnt_exist.txt",
  s3_path("s64-test-2", "file_doesnt_exist.txt")
)
\dontshow{\}) # examplesIf}
}
