% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{aws_group}
\alias{aws_group}
\title{Get a group}
\usage{
aws_group(name)
}
\arguments{
\item{name}{(character) the group name}
}
\value{
A list with two elements:
\itemize{
\item group: tibble with information about the group
\item users: tibble with users in the group
}
}
\description{
Get a group
}
\details{
see docs \url{https://www.paws-r-sdk.com/docs/iam_get_group/}
}
\examples{
\dontrun{
aws_group(name="users")
}
}
