% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{aws_group_create}
\alias{aws_group_create}
\title{Create a group}
\usage{
aws_group_create(name, path = NULL)
}
\arguments{
\item{name}{(character) A group name. required}

\item{path}{(character) The path for the group name. optional.
If it is not included, it defaults to a slash (/).}
}
\value{
A tibble with information about the group created
}
\description{
Create a group
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_create_group/}
docs for details on the parameters
}
\examples{
\dontshow{if (aws_has_creds()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_group_create("testingagroup")
aws_group("testingagroup")
# cleanup
aws_group_delete("testingagroup")
\dontshow{\}) # examplesIf}
}
\seealso{
Other groups: 
\code{\link{aws_group}()},
\code{\link{aws_group_delete}()},
\code{\link{aws_group_exists}()},
\code{\link{aws_groups}()},
\code{\link{six_group_delete}()}
}
\concept{groups}
