% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{aws_group_delete}
\alias{aws_group_delete}
\title{Delete a group}
\usage{
aws_group_delete(name)
}
\arguments{
\item{name}{(character) A group name. required}
}
\value{
an empty list
}
\description{
Delete a group
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_group/}
docs for more details
}
\examples{
\dontshow{if (aws_has_creds()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_group_create("somegroup")
aws_group_delete("somegroup")
\dontshow{\}) # examplesIf}
}
\seealso{
Other groups: 
\code{\link{aws_group}()},
\code{\link{aws_group_create}()},
\code{\link{aws_group_exists}()},
\code{\link{aws_groups}()}
}
\concept{groups}
