% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{aws_policy_create}
\alias{aws_policy_create}
\title{Create a policy}
\usage{
aws_policy_create(name, document, path = NULL, description = NULL, tags = NULL)
}
\arguments{
\item{name}{(character) a policy name. required}

\item{document}{(character) the policy document you want to use
as the content for the new policy. required.}

\item{path}{(character) the path for the policy. if not given
default is "/". optional}

\item{description}{(character) a friendly description of the policy.
optional. cannot be changed after assigning it}

\item{tags}{(character) a vector of tags that you want to attach to
the new IAM policy. Each tag consists of a key name and an associated
value. optional}
}
\value{
a tibble with policy details
}
\description{
Create a policy
}
\details{
see docs \url{https://www.paws-r-sdk.com/docs/iam_create_policy/}
}
\examples{
\dontrun{
aws_db_rds_list()
aws_policy_document_create()
aws_policy_create("RdsAllow", document = doc)
}
}
\seealso{
Other policies: 
\code{\link{as_policy_arn}()},
\code{\link{aws_policies}()},
\code{\link{aws_policy_attach}()},
\code{\link{aws_policy_detach}()},
\code{\link{aws_policy_exists}()},
\code{\link{aws_policy}()}
}
\concept{policies}
