% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{aws_policy_delete_version}
\alias{aws_policy_delete_version}
\title{Delete a policy version}
\usage{
aws_policy_delete_version(name, version_id)
}
\arguments{
\item{name}{(character) a policy name. required. within the function
we lookup the policy arn which is what's passed to the AWS API}

\item{version_id}{(character) The policy version to delete. required.
Allows (via regex) a string of characters that consists of the lowercase
letter 'v' followed by one or two digits, and optionally followed by a
period '.' and a string of letters and digits.}
}
\value{
invisibly returns nothing
}
\description{
Delete a policy version
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
doc <- aws_policy_document_create(
  region = "us-east-2",
  account_id = "1234567890",
  resource_id = "*",
  user = "jane_doe",
  action = "rds-db:connect"
)
aws_policy_create("RdsAllow456", document = doc)
aws_policy_delete_version("RdsAllow456", "v1")
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.paws-r-sdk.com/docs/iam_delete_policy_version/}
}
\seealso{
Other policies: 
\code{\link{as_policy_arn}()},
\code{\link{aws_policies}()},
\code{\link{aws_policy}()},
\code{\link{aws_policy_attach}()},
\code{\link{aws_policy_create}()},
\code{\link{aws_policy_delete}()},
\code{\link{aws_policy_detach}()},
\code{\link{aws_policy_exists}()},
\code{\link{aws_policy_list_entities}()},
\code{\link{aws_policy_list_versions}()}
}
\concept{policies}
