% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{aws_policy_document_create}
\alias{aws_policy_document_create}
\title{Create a policy document}
\usage{
aws_policy_document_create(
  region,
  account_id,
  resource_id,
  user,
  effect = "Allow",
  ...
)
}
\arguments{
\item{region}{(character) the AWS Region for the DB instance. length==1}

\item{account_id}{(character) the AWS account number for the DB instance.
length==1. The user must be in the same account as the account for the
DB instance}

\item{resource_id}{(character) the identifier for the DB instance. length==1}

\item{user}{(character) a user name that has an IAM account. length>=1}

\item{effect}{(character) valid values: "Allow" (default), "Deny". length==1}

\item{...}{named args passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}}
}
\value{
a json class string. use \code{\link[=as.character]{as.character()}} to coerce to a regular
string
}
\description{
Create a policy document
}
\note{
a few document items are hard-coded:
\itemize{
\item \code{Version} is set to 2012-10-17"
\item \code{Action} is set to "rds-db:connect"
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
### DB account = user in a database that has access to it
# all DB instances & DB accounts for a AWS account and AWS Region
aws_policy_document_create("us-east-2", "1234567890", "*", "*")
# all DB instances for a AWS account and AWS Region, single DB account
aws_policy_document_create("us-east-2", "1234567890", "*", "jane_doe")
# single DB instasnce, single DB account
aws_policy_document_create(
  "us-east-2",
  "1234567890", "db-ABCDEFGHIJKL01234", "jane_doe"
)
# single DB instance, many users
aws_policy_document_create(
  "us-east-2", "1234567890",
  "db-ABCDEFGHIJKL01234", c("jane_doe", "mary_roe")
)
\dontshow{\}) # examplesIf}
}
\references{
#no lint start
https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_elements.html
#no lint end
}
