% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{aws_policy_exists}
\alias{aws_policy_exists}
\title{Check if a policy exists}
\usage{
aws_policy_exists(arn)
}
\value{
single logical, \code{TRUE} or \code{FALSE}
}
\description{
Checks for both customer managed and AWS managed policies
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_policy_exists("ReadOnlyAccess")
aws_policy_exists("arn:aws:iam::aws:policy/ReadOnlyAccess")
\dontshow{\}) # examplesIf}
}
\seealso{
Other policies: 
\code{\link{as_policy_arn}()},
\code{\link{aws_policies}()},
\code{\link{aws_policy_attach}()},
\code{\link{aws_policy_create}()},
\code{\link{aws_policy_delete_version}()},
\code{\link{aws_policy_delete}()},
\code{\link{aws_policy_detach}()},
\code{\link{aws_policy_list_entities}()},
\code{\link{aws_policy_list_versions}()},
\code{\link{aws_policy}()}
}
\concept{policies}
