% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{aws_role}
\alias{aws_role}
\title{Get a role}
\usage{
aws_role(name)
}
\arguments{
\item{name}{(character) the role name}
}
\value{
a tibble with role details
}
\description{
Get a role
}
\details{
see docs \url{https://www.paws-r-sdk.com/docs/iam_get_role/}
}
\examples{
\dontrun{
aws_role(name = "OrganizationAccountSecurityRole")
}
}
