% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{aws_role_delete}
\alias{aws_role_delete}
\title{Delete a role}
\usage{
aws_role_delete(name)
}
\arguments{
\item{name}{(character) A role name. required}
}
\value{
\code{NULL} invisibly
}
\description{
Delete a role
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_role/}
docs for more details
}
\examples{
\dontshow{if (aws_has_creds()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
if (aws_role_exists(name = "MyRole")) {
  aws_role_delete(name = "MyRole")
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other roles: 
\code{\link{aws_role}()},
\code{\link{aws_role_create}()},
\code{\link{aws_role_exists}()},
\code{\link{aws_roles}()}
}
\concept{roles}
