% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{aws_role_delete}
\alias{aws_role_delete}
\title{Delete a role}
\usage{
aws_role_delete(name)
}
\arguments{
\item{name}{(character) A role name. required}
}
\value{
an empty list
}
\description{
Delete a role
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_role/}
docs for more details
}
\examples{
\dontrun{
aws_role_delete(name = "MyRole")
}
}
