% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{aws_role_exists}
\alias{aws_role_exists}
\title{Check if a role exists}
\usage{
aws_role_exists(name)
}
\arguments{
\item{name}{(character) the role name}
}
\value{
a single boolean
}
\description{
Check if a role exists
}
\seealso{
Other roles: 
\code{\link{aws_role_create}()},
\code{\link{aws_role_delete}()},
\code{\link{aws_roles}()},
\code{\link{aws_role}()}
}
\concept{roles}
