% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{aws_roles}
\alias{aws_roles}
\title{List roles}
\usage{
aws_roles(...)
}
\arguments{
\item{...}{parameters passed on to the \code{paws}
\href{https://www.paws-r-sdk.com/docs/iam_list_roles/}{list_users} method}
}
\value{
A tibble with information about roles
}
\description{
List roles
}
\examples{
\dontshow{if (interactive() && aws_has_creds()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_roles()
\dontshow{\}) # examplesIf}
}
\seealso{
Other roles: 
\code{\link{aws_role}()},
\code{\link{aws_role_create}()},
\code{\link{aws_role_delete}()},
\code{\link{aws_role_exists}()}
}
\concept{roles}
