% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{aws_s3_policy_doc_create}
\alias{aws_s3_policy_doc_create}
\title{Create a policy document for an S3 bucket}
\usage{
aws_s3_policy_doc_create(bucket, action, effect = "Allow", sid = NULL, ...)
}
\arguments{
\item{bucket}{(character) bucket name. required}

\item{action}{(character) an action. required. see Actions below.}

\item{effect}{(character) valid values: "Allow" (default), "Deny". length==1}

\item{sid}{(character) a statement id. optional}

\item{...}{Additional named arguments. See link in Details for options,
and examples below}
}
\value{
a policy document as JSON (of class \code{json})
}
\description{
Create a policy document for an S3 bucket
}
\details{
There's this separate function for creating policy docs for S3 because
buckets are globally unique, so AWS figures out the region and account
ID for you.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_s3_policy_doc_create(
  bucket = "s64-test-22",
  action = s3_actions_read()
)
\dontshow{\}) # examplesIf}
}
