% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{aws_secrets_delete}
\alias{aws_secrets_delete}
\title{Delete a secret}
\usage{
aws_secrets_delete(id, ...)
}
\arguments{
\item{id}{(character) The name or ARN of the secret. required}

\item{...}{further named parameters passed on to \code{delete_secret}
\url{https://www.paws-r-sdk.com/docs/secretsmanager_delete_secret/}}
}
\value{
(list) with fields:
\itemize{
\item ARN
\item Name
\item DeletionDate
}
}
\description{
Delete a secret
}
\examples{
\dontrun{
# Does exist
aws_secrets_delete(id = "MyTestDatabaseSecret")

# Does not exist
# aws_secrets_get(id = "DoesntExist")
#> Error: ResourceNotFoundException (HTTP 400). Secrets Manager
#>   can't find the specified secret.
}
}
