% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{aws_secrets_list}
\alias{aws_secrets_list}
\title{List secrets}
\usage{
aws_secrets_list()
}
\value{
(list) list with secrets
}
\description{
List secrets
}
\examples{
\dontrun{
aws_secrets_list()
}
}
