% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets_manager.R
\name{aws_secrets_pwd}
\alias{aws_secrets_pwd}
\title{Get a random password}
\usage{
aws_secrets_pwd(...)
}
\arguments{
\item{...}{named parameters passed on to \code{get_random_password}
\url{https://www.paws-r-sdk.com/docs/secretsmanager_get_random_password/}}
}
\description{
Get a random password
}
\examples{
\dontrun{
aws_secrets_pwd()
aws_secrets_pwd(ExcludeNumbers = TRUE)
}
}
