% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user}
\alias{aws_user}
\title{Get a user}
\usage{
aws_user(username = NULL)
}
\arguments{
\item{username}{(character) A user name. required}
}
\value{
a named list with slots for:
\itemize{
\item user (tibble)
\item policies (list)
\item groups (list)
\item attached_policies (list)
}
}
\description{
Gets user information, including policies, groups, and attached policies
}
\details{
See the following docs links for details
\itemize{
\item \url{https://www.paws-r-sdk.com/docs/iam_get_user/}
\item \url{https://www.paws-r-sdk.com/docs/iam_list_user_policies/}
\item \url{https://www.paws-r-sdk.com/docs/iam_list_groups_for_user/}
\item \url{https://www.paws-r-sdk.com/docs/iam_list_attached_user_policies/}
}
}
\examples{
\dontrun{
# if username not supplied, gets logged in user
aws_user()

# supply a username to get that user's information
aws_user_create("testBlueBird")
aws_user(username = "testBlueBird")
aws_user_delete("testBlueBird") # cleanup user
}
}
