% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_access_key}
\alias{aws_user_access_key}
\title{Get AWS Access Key for a user}
\usage{
aws_user_access_key(username = NULL, ...)
}
\arguments{
\item{username}{(character) A user name. required}

\item{...}{further named args passed on to
\href{https://www.paws-r-sdk.com/docs/iam_list_access_keys/}{list_access_keys}}
}
\value{
a tibble with key details
}
\description{
IMPORTANT: the secret access key is only accessible during key
and user creation
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_list_access_keys/}
docs for more details
}
\seealso{
Other users: 
\code{\link{aws_user_access_key_delete}()},
\code{\link{aws_user_add_to_group}()},
\code{\link{aws_user_create}()},
\code{\link{aws_user_current}()},
\code{\link{aws_user_delete}()},
\code{\link{aws_user_exists}()},
\code{\link{aws_users}()},
\code{\link{aws_user}()},
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{users}
