% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_access_key}
\alias{aws_user_access_key}
\title{Get the current user's AWS Access Key}
\usage{
aws_user_access_key()
}
\value{
a tibble with key details
}
\description{
IMPORTANT: the secret access key is only accessible during key
and user creation
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_list_access_keys/}
docs for more details
}
\examples{
aws_user_access_key()
}
