% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_access_key_delete}
\alias{aws_user_access_key_delete}
\title{Delete current user's AWS Access Key}
\usage{
aws_user_access_key_delete(access_key_id, username = NULL)
}
\arguments{
\item{access_key_id}{(character) The access key ID for the access key ID
and secret access key you want to delete. required.}

\item{username}{(character) A user name. optional. however, if you do
not supply a username, \code{paws} will likely use the current user, and so
may not be the user the access key id is associated - and then you'll get
an error like \verb{NoSuchEntity (HTTP 404). The Access Key with id xx cannot be found}}
}
\value{
NULL, invisibly
}
\description{
Delete current user's AWS Access Key
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_access_key/}
docs for more details
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_user_access_key_delete(access_key_id = "adfasdfadfadfasdf")
aws_user_access_key_delete(access_key_id = "adfasdf", username = "jane")
\dontshow{\}) # examplesIf}
}
\seealso{
Other users: 
\code{\link{aws_user_access_key}()},
\code{\link{aws_user_add_to_group}()},
\code{\link{aws_user_create}()},
\code{\link{aws_user_current}()},
\code{\link{aws_user_delete}()},
\code{\link{aws_user_exists}()},
\code{\link{aws_users}()},
\code{\link{aws_user}()},
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{users}
