% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_add_to_group}
\alias{aws_user_add_to_group}
\title{Add a user to a group}
\usage{
aws_user_add_to_group(username, groupname)
}
\arguments{
\item{username}{(character) A user name. required}

\item{groupname}{(character) a group name. required}
}
\value{
a named list with slots for:
\itemize{
\item user (tibble)
\item policies (list)
\item attached_policies (list)
\item groups (list)
}
}
\description{
Add a user to a group
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_add_user_to_group/}
docs for more details
}
\examples{
\dontrun{
if (!aws_group_exists("testgroup3")) {
  aws_group_create("testgroup3")
}
if (!aws_user_exists("testBlueBird3")) {
  aws_user_create("testBlueBird3")
}
aws_user_add_to_group(username = "testBlueBird3", groupname = "testgroup3")
}
}
\seealso{
Other users: 
\code{\link{aws_user_access_key_delete}()},
\code{\link{aws_user_access_key}()},
\code{\link{aws_user_create}()},
\code{\link{aws_user_current}()},
\code{\link{aws_user_delete}()},
\code{\link{aws_user_exists}()},
\code{\link{aws_users}()},
\code{\link{aws_user}()}
}
\concept{users}
