% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_current}
\alias{aws_user_current}
\title{Get the current logged-in username as a string}
\usage{
aws_user_current()
}
\value{
username as character
}
\description{
Get the current logged-in username as a string
}
