% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_current}
\alias{aws_user_current}
\title{Get the current logged-in username as a string}
\usage{
aws_user_current()
}
\value{
username as character, scalar
}
\description{
Get the current logged-in username as a string
}
\seealso{
Other users: 
\code{\link{aws_user}()},
\code{\link{aws_user_access_key}()},
\code{\link{aws_user_access_key_delete}()},
\code{\link{aws_user_add_to_group}()},
\code{\link{aws_user_create}()},
\code{\link{aws_user_delete}()},
\code{\link{aws_user_exists}()},
\code{\link{aws_users}()},
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{users}
