% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_delete}
\alias{aws_user_delete}
\title{Delete a user}
\usage{
aws_user_delete(username)
}
\arguments{
\item{username}{(character) A user name. required}
}
\value{
an empty list
}
\description{
Delete a user
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_user/}
docs for more details
}
\examples{
\dontrun{
aws_user_delete(username = "testBlueBird")
}
}
