% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_delete}
\alias{aws_user_delete}
\title{Delete a user}
\usage{
aws_user_delete(username)
}
\arguments{
\item{username}{(character) A user name. required}
}
\value{
NULL invisibly
}
\description{
Delete a user
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_user/}
docs for more details
}
\examples{
\dontshow{if (interactive() && aws_has_creds()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
user_name <- random_user()
aws_user_create(user_name)
aws_user_delete(user_name)
aws_user_exists(user_name)
\dontshow{\}) # examplesIf}
}
\seealso{
Other users: 
\code{\link{aws_user}()},
\code{\link{aws_user_access_key}()},
\code{\link{aws_user_access_key_delete}()},
\code{\link{aws_user_add_to_group}()},
\code{\link{aws_user_create}()},
\code{\link{aws_user_current}()},
\code{\link{aws_user_exists}()},
\code{\link{aws_users}()},
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{users}
