% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_user_delete}
\alias{aws_user_delete}
\title{Delete a user}
\usage{
aws_user_delete(username)
}
\arguments{
\item{username}{(character) A user name. required}
}
\value{
\code{NULL} invisibly
}
\description{
Delete a user
}
\details{
See \url{https://www.paws-r-sdk.com/docs/iam_delete_user/}
docs for more details
}
\examples{
\dontrun{
aws_user_delete(username = "testBlueBird")
}
}
\seealso{
Other users: 
\code{\link{aws_user}()},
\code{\link{aws_user_access_key}()},
\code{\link{aws_user_access_key_delete}()},
\code{\link{aws_user_add_to_group}()},
\code{\link{aws_user_create}()},
\code{\link{aws_user_current}()},
\code{\link{aws_user_exists}()},
\code{\link{aws_users}()},
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{users}
