% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_users}
\alias{aws_users}
\title{List Users}
\usage{
aws_users(...)
}
\arguments{
\item{...}{parameters passed on to the \code{paws}
\href{https://www.paws-r-sdk.com/docs/iam_list_users/}{list_users} method}
}
\value{
A tibble with information about user accounts
}
\description{
List Users
}
\examples{
\dontrun{
aws_users()
}
}
