% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{aws_users}
\alias{aws_users}
\title{List Users}
\usage{
aws_users(con = NULL, ...)
}
\arguments{
\item{...}{parameters passed on to the \code{paws}
\href{https://www.paws-r-sdk.com/docs/iam_list_users/}{list_users} method}
}
\value{
A tibble with information about user accounts
}
\description{
List Users
}
\examples{
\dontrun{
aws_users()
}
}
\seealso{
Other users: 
\code{\link{aws_user_access_key}()},
\code{\link{aws_user_add_to_group}()},
\code{\link{aws_user_create}()},
\code{\link{aws_user_current}()},
\code{\link{aws_user_delete}()},
\code{\link{aws_user_exists}()},
\code{\link{aws_user}()}
}
\concept{users}
