% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{aws_vpc_security_group}
\alias{aws_vpc_security_group}
\title{Get a security group by ID}
\usage{
aws_vpc_security_group(id, ...)
}
\arguments{
\item{id}{(character) The id of the security group. required}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/ec2_describe_security_groups/}{describe_security_groups}}
}
\value{
(list) with fields:
\itemize{
\item SecurityGroups (list) each security group
\itemize{
\item Description
\item GroupName
\item IpPermissions
\item OwnerId
\item GroupId
\item IpPermissionsEgress
\item Tags
\item VpcId
}
\item NextToken (character) token for paginating
}
}
\description{
Get a security group by ID
}
\seealso{
Other security groups: 
\code{\link{aws_vpc_sec_group_rules}()},
\code{\link{aws_vpc_security_group_create}()},
\code{\link{aws_vpc_security_group_ingress}()},
\code{\link{aws_vpc_security_groups}()},
\code{\link{aws_vpc_sg_with_ingresss}()},
\code{\link{con_ec2}()}
}
\concept{security groups}
