% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{aws_vpc_security_group_create}
\alias{aws_vpc_security_group_create}
\title{Create a security group}
\usage{
aws_vpc_security_group_create(
  name,
  engine = "mariadb",
  description = NULL,
  vpc_id = NULL,
  tags = NULL,
  ...
)
}
\arguments{
\item{name}{(character) The name of the new secret. required}

\item{engine}{(character) The engine to use. default: "mariadb". required.
one of: mariadb, mysql, or postgres}

\item{description}{(character) The description of the secret. optional}

\item{vpc_id}{(character) a VPC id. optional. if not supplied your default
VPC is used. To get your VPCs, see \code{\link[=aws_vpcs]{aws_vpcs()}}}

\item{tags}{(character) The tags to assign to the security group. optional}

\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/ec2_create_security_group/}{create_security_group}}
}
\value{
(list) with fields:
\itemize{
\item GroupId (character)
\item Tags (list)
}
}
\description{
Create a security group
}
\examples{
\dontrun{
# create security group
x <- aws_vpc_security_group_create(
  name = "testing1",
  description = "Testing security group creation"
)
aws_vpc_security_group_create(name = "testing2")
aws_vpc_security_group_create(
  name = "testing4",
  tags = list(
    list(
      ResourceType = "security-group",
      Tags = list(
        list(
          Key = "sky",
          Value = "blue"
        )
      )
    )
  )
)

# add ingress
aws_vpc_security_group_ingress(
  id = x$GroupId,
  ip_permissions = ip_permissions_generator("mariadb")
)
}
}
\seealso{
Other security groups: 
\code{\link{aws_vpc_sec_group_rules}()},
\code{\link{aws_vpc_security_group_ingress}()},
\code{\link{aws_vpc_security_groups}()},
\code{\link{aws_vpc_security_group}()},
\code{\link{aws_vpc_sg_with_ingress}()}
}
\concept{security groups}
