% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_security_groups.R
\name{aws_vpc_security_groups}
\alias{aws_vpc_security_groups}
\title{List VPC security groups}
\usage{
aws_vpc_security_groups(...)
}
\arguments{
\item{...}{named parameters passed on to \href{https://www.paws-r-sdk.com/docs/ec2_describe_security_groups/}{describe_security_groups}}
}
\value{
(list) list with security groups
}
\description{
List VPC security groups
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aws_vpc_security_groups()
aws_vpc_security_groups(MaxResults = 6)
\dontshow{\}) # examplesIf}
}
\seealso{
Other security groups: 
\code{\link{aws_vpc_sec_group_rules}()},
\code{\link{aws_vpc_security_group_create}()},
\code{\link{aws_vpc_security_group_ingress}()},
\code{\link{aws_vpc_security_group}()},
\code{\link{aws_vpc_sg_with_ingress}()}
}
\concept{security groups}
