% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{cluster_details}
\alias{cluster_details}
\title{Get information for all clusters}
\usage{
cluster_details()
}
\value{
a list of cluster details
}
\description{
Get information for all clusters
}
\keyword{internal}
