% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clients.R
\name{con_factory}
\alias{con_factory}
\title{\code{paws} connection factory}
\usage{
con_factory(service)
}
\arguments{
\item{service}{(character) any service supported by \pkg{paws}}
}
\value{
a function for returning the client for the service in
the \code{service} parameter
}
\description{
\code{paws} connection factory
}
\keyword{internal}
