% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam.R
\name{con_iam}
\alias{con_iam}
\title{Get the \code{paws} Identity and Access Management (IAM) client}
\usage{
con_iam()
}
\value{
a list with methods for interfacing with IAM;
see \url{https://www.paws-r-sdk.com/docs/iam/}
}
\description{
Get the \code{paws} Identity and Access Management (IAM) client
}
\keyword{internal}
