% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3con.R
\name{con_s3fs}
\alias{con_s3fs}
\title{s3fs connection}
\usage{
con_s3fs()
}
\description{
s3fs connection
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
con <- con_s3fs()
file_copy <- con_s3fs()$file_copy
\dontshow{\}) # examplesIf}
}
