% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{figure_out_policy_arn}
\alias{figure_out_policy_arn}
\title{Figure out policy Arn from a name}
\usage{
figure_out_policy_arn(name)
}
\description{
Figure out policy Arn from a name
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# user managed, exists
figure_out_policy_arn("MyTestPolicy")
# user managed, doesn't exist
figure_out_policy_arn("DoesNotExist")
# aws managed
figure_out_policy_arn("AmazonS3ReadOnlyAccess")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
