% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paginate_aws_token}
\alias{paginate_aws_token}
\title{Paginate over list_* methods with NextToken}
\usage{
paginate_aws_token(fun, target, ...)
}
\arguments{
\item{fun}{(function) a function to call}

\item{target}{(character) a list element to get}

\item{...}{named args passed on to \code{fun}}
}
\description{
Paginate over list_* methods with NextToken
}
\examples{
\dontrun{
paginate_aws_token(
  fun = env64$secretsmanager$list_secrets,
  target = "SecretList"
)
}
}
\keyword{internal}
