% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_s3_parse}
\alias{path_s3_parse}
\title{Parse s3 paths}
\usage{
path_s3_parse(paths)
}
\arguments{
\item{paths}{(character) one or more s3 paths}
}
\value{
an unnamed list with each slot a named list with bucket, path,
and file
}
\description{
Parse s3 paths
}
\keyword{internal}
