% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random_string}
\alias{random_string}
\title{Get a random string with prefix}
\usage{
random_string(prefix, size = 8)
}
\arguments{
\item{prefix}{(character) any string. required.}

\item{size}{(character) length of the random part (not including
\code{prefix})}
}
\value{
(character) a string with \code{prefix} at beginning
}
\description{
Get a random string with prefix
}
\examples{
random_string("bucket")
replicate(10, random_string("bucket"))
}
