% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_user.R
\name{random_user}
\alias{random_user}
\title{Get a random user}
\usage{
random_user()
}
\value{
(character) a username with a random adjective plus a
random noun combined into one string, shortened to no longer than 16
characters if longer than 16
}
\description{
Get a random user
}
\examples{
random_user()
replicate(10, random_user())
}
