% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-s3fs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{s3_path}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{s3fs}{\code{\link[s3fs:path]{s3_path}}}
}}

