% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{six_admin_setup}
\alias{six_admin_setup}
\title{AWS account setup for administrators}
\usage{
six_admin_setup()
}
\value{
NULL invisibly
}
\description{
AWS account setup for administrators
}
\section{What is magical}{

\itemize{
\item Setup a users IAM group: users that do not require admin persmissions
\item Add policies to the users group
\item Setup an admin IAM group: users that require admin permissions
\item Add policies to the admin group
\item (NOT WORKING YET) Check that a user can be added to a group and has
access as expected
}
}

\seealso{
Other magicians: 
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{magicians}
