% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{six_admin_setup}
\alias{six_admin_setup}
\title{AWS account setup for administrators}
\usage{
six_admin_setup(users_group = "users", admin_group = "admin")
}
\arguments{
\item{users_group}{(character) name for the users group. default: "users"}

\item{admin_group}{(character) name for the admin group. default: "admin"}
}
\value{
NULL invisibly
}
\description{
AWS account setup for administrators
}
\section{What is magical}{

\itemize{
\item Setup a users IAM group: users that do not require admin persmissions
\item Add policies to the users group
\item Setup an admin IAM group: users that require admin permissions
\item Add policies to the admin group
}
}

\seealso{
Other magicians: 
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{magicians}
