% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{six_bucket_upload}
\alias{six_bucket_upload}
\title{Magically upload a mix of files and directories into a bucket}
\usage{
six_bucket_upload(path, remote, force = FALSE, ...)
}
\arguments{
\item{path}{(character) one or more file paths to add to
the \code{bucket}. required. can include directories or files}

\item{remote}{(character) a vector of paths to use to upload
files in \code{path}. the bucket name is determined from this param.
if the vector passed in is length>1 then the base path component
must all be the same. required}

\item{force}{(logical) force bucket creation without going through
the prompt. default: \code{FALSE}. Should only be set to \code{TRUE} when
required for non-interactive use.}

\item{...}{named params passed on to
\href{https://www.paws-r-sdk.com/docs/s3_put_object/}{put_object}}
}
\value{
(character) a vector of remote s3 paths where your
files are located
}
\description{
Magically upload a mix of files and directories into a bucket
}
\section{What is magical}{

\itemize{
\item Exits early if folder or files do not exist
\item Creates the bucket if it does not exist
\item Adds files to the bucket, figuring out the key to use from
the supplied path
\item Function is vectoried for the \code{path} argument; you can
pass in many paths
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# single file, single remote path
bucket <- random_string("bucket")
demo_rds_file <- file.path(system.file(), "Meta/demo.rds")
six_bucket_upload(path = demo_rds_file, remote = bucket)

## a file and a directory - with a single remote path
bucket <- random_string("bucket")
library(fs)
tdir <- path(path_temp(), "mytmp")
dir_create(tdir)
purrr::map(letters, \(l) file_create(path(tdir, l)))
six_bucket_upload(path = c(demo_rds_file, tdir), remote = bucket)

## two files - two values passed to remote path
bucket <- random_string("bucket")
links_file <- file.path(system.file(), "Meta/links.rds")
six_bucket_upload(path = c(demo_rds_file, links_file),
 remote = path(bucket, c("afile.txt", "anotherfile.txt")))
\dontshow{\}) # examplesIf}
}
\seealso{
Other buckets: 
\code{\link{aws_bucket_create}()},
\code{\link{aws_bucket_delete}()},
\code{\link{aws_bucket_download}()},
\code{\link{aws_bucket_exists}()},
\code{\link{aws_bucket_list_objects}()},
\code{\link{aws_bucket_tree}()},
\code{\link{aws_bucket_upload}()},
\code{\link{aws_buckets}()},
\code{\link{six_bucket_delete}()}

Other magicians: 
\code{\link{six_admin_setup}()},
\code{\link{six_bucket_delete}()},
\code{\link{six_file_upload}()},
\code{\link{six_user_create}()},
\code{\link{six_user_delete}()}
}
\concept{buckets}
\concept{magicians}
