% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixtyfour-package.R
\docType{package}
\name{sixtyfour-package}
\alias{sixtyfour}
\alias{sixtyfour-package}
\title{sixtyfour: Humane Interface to AWS}
\description{
Humane Interface to AWS. And more text. Even more text.
}
\author{
\strong{Maintainer}: Sean Kross \email{skross@fredhutch.org}

Authors:
\itemize{
  \item Scott Chamberlain
}

}
\keyword{internal}
