% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixtyfour-package.R
\docType{package}
\name{sixtyfour-package}
\alias{sixtyfour}
\alias{sixtyfour-package}
\title{sixtyfour: Humane Interface to AWS}
\description{
Write some more things here to give folks a
sense of what the goal of this package is,
high level entrypoints, etc.
}
\author{
\strong{Maintainer}: Scott Chamberlain \email{sachamber@fredhutch.org}

Authors:
\itemize{
  \item Sean Kross
}

}
\keyword{internal}
