% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixtyfour-package.R
\docType{package}
\name{sixtyfour-package}
\alias{sixtyfour}
\alias{sixtyfour-package}
\title{sixtyfour: Humane Interface to AWS}
\description{
An opinionated interface to AWS, with methods for interacting with 'IAM', S3, RDS, Redshift, and Billing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/getwilds/sixtyfour}
  \item \url{https://getwilds.org/sixtyfour/}
  \item Report bugs at \url{https://github.com/getwilds/sixtyfour/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{sachamber@fredhutch.org} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})

Authors:
\itemize{
  \item Sean Kross
}

Other contributors:
\itemize{
  \item Fred Hutchinson Cancer Center \email{wilds@fredhutch.org} [funder]
}

}
\keyword{internal}
