% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fetch_secret.R
\name{ui_fetch_secret}
\alias{ui_fetch_secret}
\title{Fetch secrets}
\usage{
ui_fetch_secret(user = NULL, password = NULL, engine = NULL, id = NULL)
}
\arguments{
\item{user}{(character) xx}

\item{password}{(character) password}

\item{engine}{(character) engine to filter secrets with. if not supplied
(\code{NULL}) all secrets are considered}

\item{id}{(character) the id of the instance. if supplied, we filter
the secret names by this id}
}
\description{
Fetch secrets
}
\section{How the function works}{

\itemize{
\item If user and password are supplied they are returned immediately
\item If user and password are not supplied, we fetch all secrets in
your secrets manager service, and then ask you which one you'd like
to use. If you choose none of them this function returns NULL for
both user and password
}
}

\keyword{internal}
