% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{with_redacted}
\alias{with_redacted}
\title{Without secrets redacted}
\usage{
with_redacted(code)
}
\arguments{
\item{code}{(expression) Code to run with secrets redacted}
}
\description{
Without secrets redacted
}
